feat: change tests structure
